+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `PsrToOrisaiClockAdapter` - makes any PSR clock compatible with `Orisai\Clock\Clock`
 - `SymfonyToOrisaiClockAdapter` - makes any symfony/clock compatible with `Orisai\Clock\Clock`
 - `ClockAdapterFactory` - chooses best supported adapter for `Orisai\Clock\Clock` compatibility
+- `OrisaiToSymfonyClockAdapter` - makes any symfony/clock compatible with `Symfony\Component\Clock\ClockInterface`
 - `FrozenClock` - accepts `DateTimeInterface` as the initial time
 
 ### Changed